several changes, incl. some for 32 bits.
In particular, while I was at it, internal output files for bw are now in
hex as they always should have been.

The ``nan'' printing bug in lingen-binary could very well be gone for
good (!)

Found also a possible off-by-one in bw/cxx/fake_fft.c

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1700 de4796e2-0b1e-0410-91b4-86bc23549317

+++ gf2x/mul-gf2x.c
@@ -121,7 +121,7 @@ void mul_gf2x_r(ulong * c,
     } else {
 	ulong *v = pool->stk + toomspace(sa);
 
-	long i, j;
+	unsigned int i, j;
 
 	memset(c, 0, sc * sizeof(ulong));
